Delete block of incorrect info from manpage
To get a solid idea of what I wanted the end-result to be with
the options and all, I decided to write the manpage before the
rest of the program.  I forgot to remove a little snippet that
dealt with using `sysexits`, however.  I may or may not
implement that in the future; I tend to like nice error codes.