Quote variables to avoid globs in configure script
If something like `!` or `*` was part of an argument to an option,
the script threw back all sorts of cryptic errors about globbing
and whatnot instead of saying "invalid feature" or whatever was
appropriate.  Quoting `$arg` seemed to fix that.

A couple trivial variables elsewhere were also wrapped in quotes.