Clean up some `#ifdef` to use `#if defined(…)`
In the case more features for `draw_penis()` come in, the system
through which what bits are included should be somewhat readable.

My only concern is cross-compiler compatibility, but I'm fairly
sure if someone's using such a queer and ancient compiler it doesn't
have `#if defined(…)` support, they can patch it to work.